requirements.txt
python-telegram-bot
psycopg2-binary
python-dotenv
opencv-python-headless
numpy
Pillow
qrcode
fastapi
python-multipart
openpyxl